move constraint lookups out of robot_configs
These depended on the hardware subpackage for some type definitions,
which won't work in robot server tests or on OT2s. Let's move them into
the hardware control backends, where they won't be loaded sometimes, and
solve the problem
